pyproject.toml
[project]
name = "grad-codex-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.13.2",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "rank-bm25>=0.2.2",
    "python-multipart>=0.0.12",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.13",
    "langchain-openai>=1.1.10",
    "langchain-classic>=1.0.1",
    "langgraph>=1.0.8",
    "matplotlib>=3.10.8",
    "python-dotenv>=1.2.1",
    "tiktoken>=0.12.0",
    # Document download scripts (scripts/)
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "playwright>=1.40.0",
    "html2text>=2024.2.26",
]
